test(interface): add exported interface test (#556)
